pyproject.toml
[project]
name = "amstar"
version = "0.1.0"
requires-python = ">=3.13,<3.14"

dependencies = [
    "alembic>=1.13.0",
    "aiomysql>=0.2.0",
    "cryptography>=42.0.0",
    "dependency-injector>=4.48.1",
    "fastapi>=0.116.0",
    "orjson>=3.10.0",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.10.0",
    "python-jose>=3.5.0",
    "ruff>=0.12.0",
    "sqlalchemy>=2.0.0",
    "typer>=0.16.0",
    "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py313"
fix = true

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
norecursedirs = [".*", "__pycache__", "venv", ".venv"]

[tool.mypy]
python_version = "3.13"
strict = false
